Make radio button 'is-selected' more permissive wrt HTML structure
With accessibility improvements to the form builder (currently in Michigan + Colorado code bases, to be pulled into styleguide), we need a more permissive search for the relevant ancestor to assign 'is-selected' to for radio buttons. This change should work for both cases: what's currently in GCF's form builer (parent), and what is in the newer version (grandparent).