Update color contrast to improve accessibility
Creates a new set of base colors for our pattern library that both has a
wider range of standard tints and shades (light and dark) so that we
don't have to manually adjust the base color each time, and has a
greater degree of contrast when used in our patterns.

In doing this, we also standardized some of the color variable names and
set up aliases to provide backward compatibility for projects that are
currently using or overriding those variables.

Also updates existing patterns for improved contrast to meet our
accessibility standards.

Signed-off-by: Christa Hartsock <christa@codeforamerica.org>
Co-authored-by: Christa Hartsock <christa@codeforamerica.org>