Update toolbar for new GCF behaviors
In GCF we want to customize the toolbar behavior on small screens so as
to maximize room for links on the right side. There are two different
behaviors specified based on the page layout:

1. On the homepage, the right-side links should wrap underneath the logo
   on the left side.
2. On form flow pages, the "GetCalFresh.org" text next to the logo icon
   should disappear. This implementation uses a `text-indent` to enable
   screen readers to still find the text.

To accomplish the behavior, this commit updates the toolbar molecule to
use flexbox layout, and adds two new variants for consumers to choose.

This also removes the "centered" variant because a brief search of
codebases known to use this style guide (GCF, BEAR) do not use this
variant.

Furthermore, this commit removes a couple non-toolbar uses of the
classes from our examples.