Fix follow-up to work for single checkbox
It looks like the follow-up question component _was_ intended to work
with checkboxes (since it clears the checked state proactively), but in
practice it didn't: When unchecking the checkbox, the call to $.show()
would trigger, resulting in the follow-up never disappearing.

By copying the conditional logic used during initialization here, we can
make it work for a single checkbox.

I've added an example to the styleguide page to assist future
implementers.