Use prism.js from cdnjs instead of a gem
The prism-rails gem is sparsley updated and includes a version of prism from 2017;
prism has been updated a number of times since then.

Additionally and more importantly, it locks the rails version at "< 6", which is not
good if you're trying to upgrade to Rails 6

The downside of loading it from a CDN is that you won't be able to view the
full-fidelity styleguide when on an airplane or exploring the earth's core.

also: change block languages from 'language-ruby' to 'language-erb' to make the syntax
highlighting more correct

Fixes #106