Adjusts how upload buttons are highlighted on focus.
The previous solution had two issues: the focus was rectangular despite the rounded rect button style, and the upload button lost its spacing, causing GCF's documents upload page to lay out incorrectly. This is a bit of a sketchy fix, since it relies on the existence of an element with class 'button' (which is actually an unfocused <label> in GCF) as a sibling of the actually-focused input element (which is invisible in GCF's case). I still need to test this in other instances of GCF's usage if there are any, but wanted to push it up to let others tell me if it's totally infeasible in the meantime.