Prevent exception when assets not precompiled
I was getting an exception when loading the styleguide documentation
after running `bin/rails s` in development because my assets weren't
precompiled, which is expected in development. Changing this
configuration

I'm not sure when this was introduced (below issue suggests 2016, so may
have appeared once we upgraded Rails/Sprockets) or if others are running into it.

More: https://github.com/rails/sprockets-rails/issues/352