Set eager_load to true in development
Recommended by the boot up message.

Also removes the disabling of the asset precompilation check, which
seemed maybe like it was addressing a different problem than intended
and possibly introducing another. The linked Github issue was talking
about performance issues, which we shouldn't run into, while we were
trying to avoid an exception (and maybe was related to the problem we've
been seeing on projects).